backend/requirements.txt
fastapi
uvicorn
pandas
yfinance
sqlalchemy
psycopg2-binary
python-dotenv
supabase
